requirements.txt
python-telegram-bot
python-dotenv
requests
magic_hour 